1. Added more tests 2. Generated strings will be escaped. 3. Better communicates are going to be thrown in case of error. 4. Unused methods removed / Some methods moved to different places.

+++ Musoq.Plugins/Lib/LibraryBaseMath.cs
@@ -55,7 +55,7 @@ namespace Musoq.Plugins
         }
 
         [BindableMethod]
-        public long? Sign(decimal? value)
+        public decimal? Sign(decimal? value)
         {
             if (!value.HasValue)
                 return null;
